backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0

# Supabase + Auth
supabase>=2.4.0
python-jose[cryptography]>=3.3.0

# HTTP
requests>=2.31.0
httpx>=0.27.0
python-multipart>=0.0.9

# Data
pandas>=2.1.0

# AI
google-genai>=0.8.0

# Legacy (remove after Phase 5 Gemini migration is verified)
openai>=1.14.0
